Refactor nzbToMediaAutoFork -> forks
