+++ tests/general.py
@@ -6,7 +6,7 @@ import requests
 
 import core
 from core import transcoder
-from core.nzbToMediaAutoFork import auto_fork
+from core.forks import auto_fork
 from core.nzbToMediaUtil import server_responding
 
 # Initialize the config
